DeepResearchFunctionApp/requirements.txt
azure-functions
azure-ai-agents>=1.1.0b3
azure-ai-projects>=1.0.0b12
azure-identity
